% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xptr.R
\name{create_xptr}
\alias{create_xptr}
\alias{null_xptr}
\title{Create an external pointer object.}
\usage{
create_xptr(address = "", tag = NULL, protected = NULL)

null_xptr(tag = NULL, protected = NULL)
}
\arguments{
\item{address}{a string of pointer address}

\item{tag}{an optional tag}

\item{protected}{an objected to be protected from gc within the lifetime of the external pointer}
}
\value{
an `externalptr` object
}
\description{
Create an external pointer object.
}
