% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xptr.R
\name{is_xpt}
\alias{is_xpt}
\title{Check if an object is an external pointer.}
\usage{
is_xpt(s)
}
\arguments{
\item{s}{an \code{externalptr} object}
}
\value{
a boolean
}
\description{
Check if an object is an external pointer.
}
