% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xptr.R
\name{register_xptr_finalizer}
\alias{register_xptr_finalizer}
\title{Register a finalizer for external poitner.}
\usage{
register_xptr_finalizer(s, f, onexit = FALSE)
}
\arguments{
\item{s}{an \code{externalptr} object}

\item{f}{an R function}

\item{onexit}{should the finalizer execute on exit?}
}
\description{
Register a finalizer for external poitner.
}
\seealso{
\code{\link[base]{reg.finalizer}}
}
