% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xptr.R
\name{set_xptr_protected}
\alias{set_xptr_protected}
\title{Set a protected R object to the external poitner.}
\usage{
set_xptr_protected(s, protected)
}
\arguments{
\item{s}{an \code{externalptr} object}

\item{protected}{an R object}
}
\description{
Set a protected R object to the external poitner.
}
