% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xptr.R
\name{set_xptr_tag}
\alias{set_xptr_tag}
\title{Set a tag to the external poitner.}
\usage{
set_xptr_tag(s, tag)
}
\arguments{
\item{s}{an `externalptr` object}

\item{tag}{an R object}
}
\description{
Set a tag to the external poitner.
}
