\name{xptr-package}
\alias{xptr-package}
\alias{xptr}
\docType{package}
\title{
\packageTitle{xptr}
}
\description{
\packageDescription{xptr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{xptr}
\packageIndices{xptr}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{xptr}

Maintainer: \packageMaintainer{xptr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
