% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xptr.R
\name{xptr_address}
\alias{xptr_address}
\title{External pointer address.}
\usage{
xptr_address(s)
}
\arguments{
\item{s}{an `externalptr` object}
}
\value{
a string of pointer address
}
\description{
External pointer address.
}
