% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xptr.R
\name{xptr_protected}
\alias{xptr_protected}
\title{Protected R object of the external pointer.}
\usage{
xptr_protected(s)
}
\arguments{
\item{s}{an `externalptr` object}
}
\value{
an R object
}
\description{
Protected R object of the external pointer.
}
