% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xptr.R
\name{xptr_tag}
\alias{xptr_tag}
\title{Tag of the external pointer.}
\usage{
xptr_tag(s)
}
\arguments{
\item{s}{An \code{externalptr} object}
}
\value{
The R object stored in the tag slot
}
\description{
Tag of the external pointer.
}
