% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xptr.R
\name{is_null_xptr}
\alias{is_null_xptr}
\title{Check if the external pointer is null.}
\usage{
is_null_xptr(s)
}
\arguments{
\item{s}{an `externalptr` object}
}
\value{
TRUE/FALSE
}
\description{
Check if the external pointer is null.
}
