% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xptr.R
\name{register_xptr_finalizer}
\alias{register_xptr_finalizer}
\title{Register a finalizer for external pointer.}
\usage{
register_xptr_finalizer(s, f, onexit = FALSE)
}
\arguments{
\item{s}{An \code{externalptr} object}

\item{f}{An R function}

\item{onexit}{should the finalizer execute on exit?}
}
\value{
No return value as the function is called for its side effect
}
\description{
Register a finalizer for external pointer.
}
\seealso{
\code{\link[base]{reg.finalizer}}
}
