% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xptr.R
\name{set_xptr_address}
\alias{set_xptr_address}
\title{Set the address of the external pointer.}
\usage{
set_xptr_address(s, p)
}
\arguments{
\item{s}{An \code{externalptr} object}

\item{p}{A string of pointer address}
}
\value{
No return value as the function is called for its side effect
}
\description{
Set the address of the external pointer.
}
