% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xptr.R
\name{set_xptr_protected}
\alias{set_xptr_protected}
\title{Set a protected R object to the external pointer.}
\usage{
set_xptr_protected(s, protected)
}
\arguments{
\item{s}{An \code{externalptr} object}

\item{protected}{An R object}
}
\value{
No return value as the function is called for its side effect
}
\description{
Set a protected R object to the external pointer.
}
