% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xptr.R
\name{set_xptr_tag}
\alias{set_xptr_tag}
\title{Set a tag to the external pointer.}
\usage{
set_xptr_tag(s, tag)
}
\arguments{
\item{s}{An \code{externalptr} object}

\item{tag}{An R object}
}
\value{
No return value as the function is called for its side effect
}
\description{
Set a tag to the external pointer.
}
