% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xptr.R
\docType{package}
\name{xptr}
\alias{xptr}
\alias{xptr-package}
\title{Manipulating External Pointer}
\description{
"xptr" provides functions to verify, create and modify "externalptr" objects.
}
\examples{
library(xptr)
a <- new_xptr("0x11a888090")
b <- new_xptr("0x11a888098")
xptr_address(b)
xptr_clear(b)
xptr_address(b)
is_null_xptr(b)
set_xptr_address(b, xptr_address(a))
xptr_address(b)
}
