% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xptr.R
\name{xptr_address}
\alias{xptr_address}
\title{External pointer address.}
\usage{
xptr_address(s, pf = TRUE)
}
\arguments{
\item{s}{An \code{externalptr} object}

\item{pf}{A \code{logical} selecting \sQuote{pointer} format (default)
or \sQuote{character} output}
}
\value{
A string of pointer address
}
\description{
External pointer address.
}
