% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xptr.R
\name{xptr_clear}
\alias{xptr_clear}
\title{Clear the pointer address of the external pointer.}
\usage{
xptr_clear(s)
}
\arguments{
\item{s}{An \code{externalptr} object}
}
\value{
No return value as the function is called for its side effect
}
\description{
Clear the pointer address of the external pointer.
}
